FederalBanner common styles
